create course component  module and handling arbitrary number of courses
