+++ courseinfo/src/components/Course.jsx
@@ -1,0 +1,45 @@
+/** @format */
+
+const Header = ({ name }) => {
+  return <h1>{name}</h1>;
+};
+
+const Content = ({ parts }) => {
+  return (
+    <div>
+      {parts.map((part) => (
+        <Part key={part.id} name={part.name} exercises={part.exercises} />
+      ))}
+    </div>
+  );
+};
+
+const Part = ({ name, exercises }) => {
+  return (
+    <p>
+      {name} {exercises}
+    </p>
+  );
+};
+
+const Total = ({ parts }) => {
+  const totalExercises = parts.reduce((acc, curr) => acc + curr.exercises, 0);
+
+  return (
+    <p>
+      Total exercises: <b>{totalExercises}</b>
+    </p>
+  );
+};
+
+const Course = ({ course }) => {
+  return (
+    <div>
+      <Header name={course.name} />
+      <Content parts={course.parts} />
+      <Total parts={course.parts} />
+    </div>
+  );
+};
+
+export default Course;
